pyproject.toml
[tool.poetry]
name = "omnixrl"
version = "0.1.0"
description = "A JAX Based Reinforcement Learning Library"
authors = ["Sandesh Katakam <sandeshkatakam@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jax = "^0.4.13"
jaxlib = "^0.4.13"
numpy = "^1.24.0"
gymnasium = "^0.29.0"
tqdm = "^4.65.0"
wandb = "^0.15.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^23.0"
isort = "^5.0"
flake8 = "^6.0"
mypy = "^1.0"
pre-commit = "^3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=omnixrl --cov-report=xml --cov-report=term-missing"